Fuse online C128 MTP plan finalization
Root cause:
The GPU MTP planner emitted batch-id placeholders, then launched the generic prefill stage-1 kernel solely to translate them into committed and speculative state slots.

Fix:
Pass request-pool indices and the state-slot offset into the MTP planner kernel so it emits final PlanC and PlanW entries in one launch. Reuse the production input types directly: int32 prefix lengths and int64 request-pool indices. Keep the generic stage-1 kernel only for ordinary prefill.

Validation:
- Online C128 MTP plan kernel test passed on GB300
- Target-verify CUDA Graph test without CPU seq_lens passed
- uvx pre-commit run --all-files --show-diff-on-failure

+++ python/sglang/test/kits/attention_unittest/attention_methods/dsv4_attention.py
@@ -630,7 +630,7 @@ def _make_forward_batch(
 ) -> ForwardBatch:
     seq_lens = case.seq_lens
     input_lens = case.input_lens
-    req_pool_indices = torch.arange(case.batch_size, dtype=torch.int32, device=device)
+    req_pool_indices = torch.arange(case.batch_size, dtype=torch.int64, device=device)
     out_cache_locs: list[int] = []
     positions: list[int] = []
